Index relations and inline tags insensitive by default.
Inline tags used to be indexed and searched case- and accent-sensitively,
but we've changed this so it defaults to insensitive instead. This seems
to be the more common use case.

Set indexing.relationsSensitive to true in blacklab[-server].yaml to keep the
old behaviour, where tags where always case-sensitive.

This setting is BlackLab-wide; there is no way to e.g. configure this per index.