Match span type by regex; implement with-spans().
with-spans(query, spansQuery, captureName) will execute the query,
capturing any spans that overlap with each hit. This makes it easy
to know what spans a hit is part of. The user might want to group
on these spans.

For example, to find any owner and/or address spans that surroung
the word 'dog':

  with-spans("dog", <'owner|address' />, 'props')

A list of spans will be captured as 'props'.

Only the first parameter is required. The second parameter defaults
to <'.+'/> (all tags); the third defaults to 'with-spans'.