Add csvdescription parameter.
The only point of this parameter is that it can be passed to BlackLab
when requesting a CSV, and if csvsummary is enabled, it will be written
to the CSV file (along with all the other summary fields). This can be
used to include a description of the search that produced these results.