Allow grouping/sorting on tag in list.
You can now group/sort on a single tag/relation from a list
captured with e.g. with-spans(...). Grouping by "listname[tagname]"
will group by the first occurrence of tagname in the list.

This is useful if you want to group/sort on any tag that occurs
in the hits found.