Fix missed parallel hits if source hits overlap.
If we have overlapping source hits, the relations we're checking against
would already be advanced too far for the next hit. The solution is
SpansBuffered, which can be marked and rewound if necessary, similar
to Java's BufferedInputStream.

Potentially useful in other cases as well; might sometimes be faster than
SpansInBuckets.