Fix old external index tags getting type 'span'.
Because of the way automatic tag capturing was implemented in the old
external index type (by inserting an implicit capture as if the user
actually wrote s:</s>), tags were captured with type "span", causing them
to be highlighted in the results. We now deal with this by explicitly
capturing them as a tag. This code path can be removed when we drop support
for the old external index (after v4 is released).