Synchronize matchInfoDefs, still issues.
The main problem is that the Spans structures can differ per segment,
based on which terms do/don't occur in that segment. If the structures differ,
the match infos captured can differ as well. But matchInfoDefs/indexes
are currently shared between all the segments.

We need to change it so each segment has its own matchInfoDefs, and they
are eventually merged when producing Hits.