Encapsulate MatchInfoDefs to fix concurrency bugs.
MatchInfoDefs used to be a list that was passed around through the
code. Because the list can change as we go through the various segments
of the Lucene index (because different parts of the query are triggered
based on which terms occur in the segment), this led to hard to debug
ConcurrentModificationExceptions and ArrayIndexOutOfBoundsExceptions.
Hopefully this is now fixed.