Smart default for sensitive= parameter.
The /termfreq and /colloc take a sensitive=true|false param
that indicates which version of the annotation to work on.

The default was false, leading to an error if the annotation
didn't have an insensitive version.

Now, if the parameter is not specified, we default to insensitive
if available, sensitive otherwise.