Make deploy work with Maven Central Publishing.
OSSRH is reaching EOL in June. This migrates the project to using
Maven Central Publishing (https://central.sonatype.org/).

A few tweaks were necessary for a clean build, such as generating
javadoc in the prepare-package phase.

For some reason, distributionManagement only seems to be required
in the blacklab-tools module (where e.g. findbugs config is stored,
and is used via a dependency).