EphemeralHit instance var naming, access.
EphemeralHit's instance variables had the same name as
the getter methods, which could be confusing.

Reduced access to package private as well, and only
used direct access to variables in HitsInternal classes.
In other classes, we use the getter methods. The JVM
should normally inline these anyway for hot code.