Remove excessive generics from Results, etc.
The Results class hierarchy and related ones like Result and ResultProperty had type arguments at the top-level. This wasn't necessary and caused problems in the subclasses.

Also introduced ResultsStatsPassive/ResultsAwaiter and used these to simplify counting in HitsFromQuery/SpansReader.

Also introduced HitsForHitProps as a limited Hits interface so that HitsInternal can be used with HitProperty directly, which will be useful for the per-segment operations we intend to implement.