Use field references instead of Strings.
A lot of places in the code still used String instead of AnnotatedField. This often meant going back and forth between the two. Now AnnotatedField is used (almost) everywhere.

This includes the BLSpans classes, which means we're deviating a bit further from Lucene's implementation, but this seems worth it in this case.