Improve how we work with MatchInfos.
MatchInfo[] may differ in size due to differences in match infos per
index segment. Due to optimization (e.g. a term might not occur in a segment),
the first segment might not register all match infos, leading to a shorter
MatchInfo array. We now use static methods in MatchInfo to deal with this
gracefully. Eventually we might want to implement a separate MatchInfos class
to enforce correct handling of match infos.

Similarly, determining match info index (groupIndex) in some HitProperty classes
needs to be done lazily because we can't predict when a match info will be
registered. Unknown match info groups will just result in an empty value,
not an error.