Reduce duplication in HitsInternal*.
Introduced HitsInternalAbstract that holds code shared between
the implementations. This also has the lock variable that is
only set for the locking implementations. Methods that are not
as performance-critical (i.e. not called for every hit) have been
moved here and check if there's a lock or not to decide what to do.