Remove captureValuePaths.
This was a little-used .blf.yaml feature to capture
a value using XPath, then substitute it into the valuePath
expression. It's better to just write an XPath expression
that handles this without the need for captureValuePaths.

The included tcf.blf.yaml format used this, but should now
work with just XPath.