HitsFromQueryAbstract. ensureResultsRead returns boolean.
HitsFromQueryAbstract is the new base class for both HitsFromQuery
and HitsFromQuerySorted (that will fetch and sort hits per segment
and then merge them; WIP).

ensureResultsRead was a void method, but it makes more sense
to return true if the requested number of results is available,
as we often check that right after calling this method.