Eliminate HitsMutable. Use HitsSimple where possible.
HitsSimple is the new name for HitsForHitProps and is a simpler
interface, focusing just on the actual hits and no other parameters
like window and sample stats, for example.

Hits[Simple].getFinishedHits() will wait until all hits are fetched
(if applicable) and then return a HitsSimple interface with the best
performance characteristics, normally a nonlocking HitsInternal
implementation. Be careful not to modify the original hits object after
calling this, though.