Perform more operations per segment. HitsResults/Hits.
Hits was renamed to HitsResults (represents everything needed
to produce a BLS response: information about the query,
window stats, sample stats, etc.).

The new Hits interface only represents a list of hits, nothing more.
HitsList* (formerly HitsInternal*) implement this interface, and
HitsFromQuery returns an implementation that lazily fetches hits as
needed.

Hits.getStatic() returns the most efficient interface for operations
that need all hits. I.e. it fetches as hits, then it might return a
nonlocking version of the locking Hits class that was used while
fetching.

HitsFromQuery now keeps hits per-segment, so we can use those for
sorting and grouping. It also maintains a global view of the segment
hits so we can return unsorted hits results.

When fetching hits, we only add hits to the global view when we find
a document boundary. That also means we may overshoot the maximum
number of hits to process/count configured a bit. As this maximum only
exists to not overstress the server, this doesn't seem like a problem.