Switch to ICU4J for Collators. CollationKey caching.
We use ICU4J for collation now. This should be more up to date
and faster than the Java Collators.

We also use collation key caching for sorting and grouping. This
is useful if many hits produce the same value, which is especially
common when grouping.

HitProperty and DocProperty were refactored to use PropContext,
which contains information about the hits and the segment (if working
per-segment), as well as the collation cache mentioned.

PropertyValueString now stores a CollationKey instead of the String.