Add run-length encodeds forward index codec.
Some annotations rarely have a value, or often have the same
value for a number of consecutive tokens. The new forward index
codec will detect this and use run-length encoding to store the
tokens. This is a trade-off: it can significantly reduce index size
(which is good for disk cache use), but decoding can be slower.

Squashed commit of the following:

commit 38be45412cdab7ad63fc2dd4e8a62a1ef1bdacee
Author: Jan Niestadt <jan.niestadt@ivdnt.org>
Date:   Mon Aug 25 16:19:28 2025 +0200

    Add test.

commit fe58e5e7db6b00e8df59727f4f1b148c7fb9d43e
Author: Jan Niestadt <jan.niestadt@ivdnt.org>
Date:   Sun Aug 24 13:28:40 2025 +0200

    Tweak.

commit f6714e613752a0c45a3f5d0ba0050d6d043cdaf0
Author: Jan Niestadt <jan.niestadt@ivdnt.org>
Date:   Sun Aug 24 11:51:08 2025 +0200

    Add run-length encoding codec for FI.

commit 397d1cdebb0474bc7b0c06b4a1cb21605a1744a2
Author: Jan Niestadt <jan.niestadt@ivdnt.org>
Date:   Sun Aug 24 10:25:39 2025 +0200

    Refactor TokensCodec into classes.