HitsFetcher fetches hits from a source (query, Hits, ...).
This is part of an effort to separate unrelated functionality
of the old HitsFromQuery class. This class performed parallel
hit fetching, but also merged these hits into a global list
or view.

If we separate the fetching from the processing of the hits, we
can reuse both the parallel fetching logic and the "global view
of segment hits" logic in other scenarios (filtering, grouping, ...).