Add run-length encodeds forward index codec.
Some annotations rarely have a value, or often have the same
value for a number of consecutive tokens. The new forward index
codec will detect this and use run-length encoding to store the
tokens. This is a trade-off: it can significantly reduce index size
(which is good for disk cache use), but decoding can be slower.

Even if it ends up a bit slower, compressible annotations are often
the more obscure ones (e.g. a certain part of speech specific
feature) which are less likely to be used to sort/group on.