Correctly keep track of stats for grouped docs.
There was a problem where a window of grouped docs would report
the total number of docs wrong (it would report the size of the window
we're looking at instead of the total number of groups found).

This makes sure to propagate the global stats (number of groups, largest group, total results in groups) properly.