Allow more advanced xpaths in annotations and metadata
Non-terminal xpaths (i.e. those that can serve as input context for subsequent xpaths, such as containerPath, or forEachPath) had to return a tangible node in the document (attribute, text or element).
This prevents some optimizations, like splitting a string right in the xpath, and indexing all sections under different subannotations.
Refactor the DocindexerSaxon to work directly with the xpath result primitives, instead of the underlying document nodes to allow this.
For some xpaths, we still need a tangible document node, because BlackLab needs to know the offsets in the document (for highlighting purposes), but this is no different than before, and you'll get the same exception if your xpath doesn't result in a node where BlackLab expects one.

In practice, this means you can now use xpaths like `tokenize(., 'some-split-char')` in the following locations where it was previously disallowed:
- (sub)annotation.basePath
- (sub)annotation.forEachPath
- metadatafield.forEachPath