Skip FileReference.DUMMY while indexing. Better indexing speed reporting.
FileIterator may produce FileReference.DUMMY whenever it encounters a file
that it decides should not be indexed. This will be detected and skipped.
This makes implementing InputSource plugins a bit easier.

When reporting indexing speed, we only start counting when we start indexing
the first actual document. This way, if it takes a few minutes to determine the
list of documents to index, the indexing speed reported won't be way off.