Use iterators in XpathFinder; avoids XdmValue wrapping
This is quite a bit faster than the somewhat more fluid XdmValue approach we were using before, though speed depends on the internal implementation in Saxon, which (currently) seems to have Iterators as the base level, with most functions wrapping the iterators. By skipping the intermediate functions we get a modest speedup.